requirements.txt
acme==3.1.0
azure-core==1.32.0
azure-identity==1.19.0
certbot==3.1.0
certbot-dns-digitalocean==3.1.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
configargparse==1.7
configobj==5.0.9
cryptography==44.0.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
filelock==3.16.1
ibm-cloud-sdk-core==3.22.1
ibm-code-engine-sdk==4.4.1
ibm-platform-services==0.59.1
identify==2.6.5
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
josepy==1.14.0
jsonpickle==4.0.1
motor==3.5.3
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
parsedatetime==2.6
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
pre-commit==4.0.1
pycparser==2.22
pydo==0.7.0
pyjwt==2.10.1
pymongo==4.8.0
pyopenssl==25.0.0
pyrfc3339==2.0.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-digitalocean==1.17.0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
six==1.17.0
tamga==0.1.8
tldextract==5.1.3
typing-extensions==4.12.2
urllib3==2.2.3
virtualenv==20.28.1
